+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - VIMEET 2110 - Add support to broadcast live video in webinar, using an iframe
 - VIMEET-2126 Add - Export all participants of sheet met
+- VIMEET-2126 - Hide participants phone / email in contact export if evaluations are below a user-defined floor
+- VIMEET-2153 - secure cookies
 
 ## [2.34.0] - 2020-06-25
 ### Added
@@ -23,7 +25,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - VIMEET-2141 - Webinar questions
 - VIMEET-2102 - Evaluate visio meeting
-- VIMEET-2126 - Hide participants phone / email in contact export if evaluations are below a user-defined floor
 - VIMEET-2128 - Show participant name on visio
 - VIMEET-2140 - Screen share without feedback
 - VIMEET-2150 - Block account in admin login after 5 failed attempts