+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [2.35.0] - 2020-06-30
 ### Added
 - VIMEET-2171 - Add participant to meeting when accessing to a videoconference meeting
 - VIMEET 2110 - Add support to broadcast live video in webinar, using an iframe